feat(core): opt-out of node:crypto, Buffer, and fileUtils
