+++ ts/packages/core/package.json
@@ -22,6 +22,18 @@
         "require": "./dist/platform/node.cjs"
       }
     },
+    "#file_tool_modifier": {
+      "workerd": "./dist/utils/modifiers/FileToolModifier.workerd.mjs",
+      "edge-light": "./dist/utils/modifiers/FileToolModifier.workerd.mjs",
+      "node": {
+        "import": "./dist/utils/modifiers/FileToolModifier.node.mjs",
+        "require": "./dist/utils/modifiers/FileToolModifier.node.cjs"
+      },
+      "default": {
+        "import": "./dist/utils/modifiers/FileToolModifier.node.mjs",
+        "require": "./dist/utils/modifiers/FileToolModifier.node.cjs"
+      }
+    },
     "#files": {
       "workerd": "./dist/models/Files.workerd.mjs",
       "edge-light": "./dist/models/Files.workerd.mjs",
