+++ ts/packages/core/src/models/Tools.ts
@@ -48,7 +48,7 @@ import {
 } from '../errors/ToolErrors';
 import { ValidationError } from '../errors/ValidationErrors';
 import { telemetry } from '../telemetry/Telemetry';
-import { FileToolModifier } from '../utils/modifiers/FileToolModifier';
+import { FileToolModifier } from '#file_tool_modifier';
 import { ComposioConfig } from '../composio';
 import { getToolkitVersion } from '../utils/toolkitVersion';
 import { handleToolExecutionError } from '../errors/ToolErrors';
